fix: multi-level nonlocal skip, import-as shadowing; split scoping tests
Two more real gaps in fact_detector_misuse.py, found by a Codex review
round on PR #929:

1. The nonlocal write-side routing fix always targeted the immediate
   lexical parent, but nonlocal can skip multiple enclosing functions --
   Python resolves it to the nearest enclosing scope that actually binds
   the name itself. Fixed by walking outward from the immediate parent,
   checking each ancestor's own locally_bound set for the declared name.

2. import json as fact / from pkg import item as fact inside a nested
   function was never recorded as a local binding at all, so it never
   shadowed an outer Fact alias. Fixed with a new ast.Import/
   ast.ImportFrom branch in the binding collector.

The four new regression tests pushed tests/test_fact_detector_misuse.py
past the architecture gate's 1200-line test-file cap (1224 lines). Split
the shadowing/scope-resolution half of TestFactEqualityMisuseSites (the
larger half after many rounds of scope-attribution findings) into a new
sibling file, tests/test_fact_detector_misuse_scoping.py -- the same
split test_mutation_run_scoping.py already made for an identical reason.
Zero behavior change: all 82 tests, same set as before, still pass.

Verified against the real repository (zero existing hits, as before),
ruff/mypy clean, architecture gate clean.
